payload: move exceptions to a separate file
